requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy
tqdm
Pillow
PyYAML
requests
gdown
git+https://github.com/openai/CLIP.git

